refactoring slt_config
