+++ setup.cfg
@@ -54,8 +54,8 @@ python =
 
 [testenv]
 allowlist_externals = pytest
-# extras =
-#    tests
+extras =
+    test
 passenv = *
 setenv =
     PYTHONPATH = {toxinidir}
