+++ tests/test_slt_from_config.py
@@ -1,91 +1,15 @@
 #! test_module_import
-
-# import itertools
-import collections
-from typing import Iterable, Generator, Dict, Union
+from nzshm_model.source_logic_tree.slt_config import (
+    get_config_groups,
+    get_config_group,
+    get_config_group_tag_permutations,
+    decompose_subduction_tag,
+    decompose_crustal_tag
+)
 from nzshm_model.source_logic_tree.logic_tree import (
-    BranchAttribute,
-    BranchAttributeValue,
     Branch,
     FaultSystemLogicTree,
 )
-
-
-def get_config_groups(logic_tree_permutations) -> Generator:
-    for permutation in logic_tree_permutations[0][0]['permute']:
-        yield permutation
-
-
-def get_config_group(logic_tree_permutations: Iterable, group_tag: str) -> Union[Dict, None]:
-    for group in get_config_groups(logic_tree_permutations):
-        if group['group'].upper() == group_tag.upper():
-            return group
-    return None
-
-
-def get_config_group_tag_permutations(logic_tree_permutations: Iterable, group_tag: str):
-    group = get_config_group(logic_tree_permutations, group_tag)
-    if group:
-        for member in group['members']:
-            yield member['tag']
-
-
-def common_tags(itm):
-    if itm[0] == 'N':
-        _n, _b = itm.split('^')
-        return BranchAttributeValue(name='bN', long_name='bN pair', value=(float(_b[1:]), float(_n[1:])))
-    if itm[0] == 'C':
-        return BranchAttributeValue(name='C', long_name='area-magnitude scaling', value=float(itm[1:]))
-    if itm[0] == 's':
-        return BranchAttributeValue(name='s', long_name='moment rate scaling', value=float(itm[1:]))
-
-
-def decompose_subduction_tag(tag):
-    """
-    "tag": "Hik TL, N16.5, b0.95, C4, s0.42",
-    "tag": "Puy 0.7, N4.6, b0.902, C4, s0.28",
-    """
-    tag = tag.replace(", b", "^b")
-    tag = tag.replace("Hik ", "Hik")
-    tag = tag.replace("Puy ", "Puy")
-    itms = tag.split(' ')
-
-    for itm in itms:
-        itm = itm.replace(',', '')  # remove commas
-        if itm[:3] in ["Hik", "Puy"]:
-            yield BranchAttributeValue(name='dm', long_name='deformation model', value=itm[3:])
-            continue
-
-        other = common_tags(itm)
-        if other:
-            yield other
-
-
-def decompose_crustal_tag(tag):
-    """
-    "tag": "geodetic, TI, N2.7, b0.823 C4.2 s0.66",
-    """
-    tag = tag.replace(", b", "^b")
-    itms = tag.split(' ')
-
-    for itm in itms:
-        itm = itm.replace(',', '')  # remove commas
-
-        if "geo" == itm[:3]:
-            yield BranchAttributeValue(
-                name='dm', long_name='deformation model', value_options=['geodetic', 'geologic'], value=itm
-            )
-            continue
-
-        if itm in 'TI, TD':
-            yield BranchAttributeValue(
-                name='td', long_name='time dependent', value_options=[True, False], value=(itm == "TD")
-            )
-            continue
-
-        other = common_tags(itm)
-        if other:
-            yield other
 
 
 class TestStructure:
@@ -123,7 +47,6 @@ class TestStructure:
         print(list(decompose_crustal_tag(group[0])))
         assert len(group) == 36
 
-
     def test_decompose_hik_tag_permutations(self):
         group = list(
             get_config_group_tag_permutations(self.model_version['model'].slt_config.logic_tree_permutations, 'HIK')
@@ -146,12 +69,18 @@ class TestStructure:
         group = get_config_group(self.model_version['model'].slt_config.logic_tree_permutations, group_key)
         print(group)
 
-
         fslt = FaultSystemLogicTree('PUY', 'Puysegur')
 
         for member in group['members']:
             bavs = list(decompose_subduction_tag(member['tag']))
-            fslt.branches.append(Branch(values=bavs, weight=member['weight'], inversion_source=member['inv_id'], distributed_source=member['bg_id']))
+            fslt.branches.append(
+                Branch(
+                    values=bavs,
+                    weight=member['weight'],
+                    inversion_source=member['inv_id'],
+                    distributed_source=member['bg_id'],
+                )
+            )
 
         print(fslt)
 
@@ -159,3 +88,4 @@ class TestStructure:
         assert fslt.branches[-1].values[1].value == (0.902, 4.6)
         assert fslt.branches[-1].values[2].value == 4.0
         assert fslt.branches[-1].values[3].value == 1.72
+        # assert 0
